68HCS12X:: add clock_t type (std::double)
git-svn-id: https://unisim.org/svn/devel@10851 ea533459-3c0a-0410-a024-a90cf4b01ff5

+++ unisim_lib/unisim/component/cxx/processor/hcs12x/types.hh
@@ -48,7 +48,7 @@ typedef uint32_t physical_address_t;
 typedef uint16_t reg_t;
 typedef uint64_t service_address_t;
 typedef uint8_t	 page_t;
-typedef uint32_t clock_t;
+typedef double clock_t;
 
 class ADDRESS {
 	public:
